+++ packages/commons/src/conditionals/validate.ts
@@ -12,14 +12,14 @@
 import Ajv from 'ajv'
 import addFormats from 'ajv-formats'
 import { ConditionalParameters, JSONSchema } from './conditionals'
-import { ErrorMapCtx, ZodIssueOptionalMessage } from 'zod'
 import { formatISO, isAfter, isBefore } from 'date-fns'
+import { ErrorMapCtx, ZodIssueOptionalMessage } from 'zod'
 import { ActionUpdate, EventState } from '../events/ActionDocument'
+import { ConditionalType, FieldConditional } from '../events/Conditional'
 import { FieldConfig } from '../events/FieldConfig'
 import { mapFieldTypeToZod } from '../events/FieldTypeMapping'
 import { FieldUpdateValue } from '../events/FieldValue'
 import { TranslationConfig } from '../events/TranslationConfig'
-import { ConditionalType, FieldConditional } from '../events/Conditional'
 
 const ajv = new Ajv({
   $data: true,
@@ -120,9 +120,9 @@ export function areConditionsMet(
   conditions: FieldConditional[],
   values: Record<string, unknown>
 ) {
-  return conditions.every((condition) =>
-    isConditionMet(condition.conditional, values)
-  )
+  return conditions.every((condition) => {
+    return ajv.validate(condition.conditional, values)
+  })
 }
 
 function isFieldConditionMet(